--- src/main/java/hana/teamfour/addminhana/controller/ExController.java
@@ -1,40 +1,0 @@
-package hana.teamfour.addminhana.controller;
-
-import javax.servlet.*;
-import javax.servlet.http.*;
-import java.io.IOException;
-import java.io.PrintWriter;
-
-public class ExController extends HttpServlet {
-    @Override
-    public void init(ServletConfig config) throws ServletException {
-        super.init(config);
-    }
-
-    @Override
-    protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-        request.setCharacterEncoding("utf-8");
-        response.setContentType("text/html;charset=utf-8");
-        PrintWriter pw = response.getWriter();
-        doHandle(request, response);
-    }
-
-    @Override
-    protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-        request.setCharacterEncoding("utf-8");
-        response.setContentType("text/html;charset=utf-8");
-        PrintWriter pw = response.getWriter();
-        doHandle(request, response);
-    }
-
-    private void doHandle(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-        request.setCharacterEncoding("utf-8");
-        response.setContentType("text/html;charset=utf-8");
-        PrintWriter pw = response.getWriter();
-    }
-
-    @Override
-    public void destroy() {
-        super.destroy();
-    }
-}
